simulate patient ae reporting test data

" "

boot_sim_test_data_patient(
  .f_sample_max_visit = function() rnorm(1, mean = 20, sd = 4),
  .f_sample_ae_per_visit = function(max_visit) rpois(max_visit, 0.5)
)

Arguments

.f_sample_max_visit
function used to sample the maximum number of aes, Default: function() rnorm(1, mean = 20, sd = 4)

.f_sample_ae_per_visit
function used to sample the aes for each visit, Default: function(x) rpois(x, 0.5)

Value

vector containing cumulative aes

Details

""

Examples

replicate(5, boot_sim_test_data_patient())

#> [[1]]
#>  [1] 0 0 0 0 0 0 1 2 3 3 3 3 3 3
#> 
#> [[2]]
#>  [1]  0  2  2  3  4  4  6  7  7  7  8  9  9  9  9 10 10 10 11
#> 
#> [[3]]
#>  [1]  0  1  1  2  3  3  4  5  8 10 10 10 10 10 10 10 11 11 12 12 12 13 13 13 13
#> [26] 13
#> 
#> [[4]]
#>  [1] 0 0 0 0 0 0 0 0 0 0 0 1 1 1 1 1
#> 
#> [[5]]
#>  [1]  2  2  2  3  3  6  6  6  7  8  8  8  9  9  9 10 11
#> 

replicate(5, boot_sim_test_data_patient(
    .f_sample_ae_per_visit = function(x) rpois(x, 1.2))
  )

#> [[1]]
#>  [1]  1  3  4  5  6  8 10 12 12 15 20 21 22 24 25 26 28 29 29
#> 
#> [[2]]
#>  [1]  0  1  3  3  3  3  5  7 10 11 12 14 14 14 17 17 22 23 23 26 26 27 27 30 32
#> 
#> [[3]]
#>  [1]  2  2  3  4  5  5  6  9 13 15 18 19 21 26 28 28 28 30 32 32 32 35
#> 
#> [[4]]
#>  [1]  2  3  4  6  8  8 10 10 12 12 13 13 15 15 18 19 20 23 25
#> 
#> [[5]]
#>  [1]  2  3  9 10 13 13 13 13 15 15 15 15 18 18
#> 

replicate(5, boot_sim_test_data_patient(
    .f_sample_max_visit = function() rnorm(1, mean = 5, sd = 5))
  )

#> [[1]]
#> [1] 1 1 3 3 4 5 6 6
#> 
#> [[2]]
#> [1] 1
#> 
#> [[3]]
#> [1] 0 2 3 4 6 6
#> 
#> [[4]]
#> [1] 0 1 1 1 2 2 4
#> 
#> [[5]]
#> [1] 0
#>