evaluate sites

evaluate site simulations by assigning a probability to each scoring values using the score distributions of the study simulations. Then calculate the P/fp ratio for each site.

eval_sites(df_sim_sites, r_sim_sites = 1000)

Arguments

df_sim_sites
dataframe generated by sim_sites()

r_sim_sites
integer, simulation repeats for prob_lower for site simulation, Default: 1000

Value

dataframe

Details

zero probabilities are being replaced by the site minimum for ecd_pval / 10 and by 1/r_sim_sites / 10

Examples

df_visit <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
    frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit$study_roche <- "A"
df_site <- site_aggr(df_visit)

df_sim_sites <- sim_sites(df_site, df_visit, r = 100)

df_eval <- eval_sites(df_sim_sites, r_sim_sites = 100)
df_eval

#> # A tibble: 5 x 16
#> # Groups:   study_roche [1]
#>   study_roche site_number visit_med75 mean_ae_site_me… mean_ae_study_m…    pval
#>   <chr>       <chr>             <dbl>            <dbl>            <dbl>   <dbl>
#> 1 A           S0001                15             3.44             6.95 1.50e-8
#> 2 A           S0002                15             3.67             6.89 2.85e-7
#> 3 A           S0003                15             7.89             5.85 1.00e+0
#> 4 A           S0004                15             7.89             5.85 1.00e+0
#> 5 A           S0005                15             8.26             5.72 1.00e+0
#> # … with 10 more variables: prob_low <dbl>, n_site <int>,
#> #   pval_n_detected <int>, pval_fp <dbl>, pval_p_vs_fp_ratio <dbl>,
#> #   pval_prob_ur <dbl>, prob_low_n_detected <int>, prob_low_fp <dbl>,
#> #   prob_low_p_vs_fp_ratio <dbl>, prob_low_prob_ur <dbl>