configure study patient pool by site parameters

for simulating a study we need to configure the study patient pool to match the configuration of the sites

get_pat_pool_config(df_visit, df_site, min_n_pat_with_med75 = 1)

Arguments

df_visit
dataframe

df_site
dataframe as created by site_aggr()

min_n_pat_with_med75
minimum number of patients with visit_med_75 for simulation, Default: 1

Value

OUTPUT_DESCRIPTION

Details

DETAILS

Examples

df_visit1 <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
                                      frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit1$study_roche <- "A"

df_visit2 <- boot_sim_test_data_study(n_pat = 1000, n_sites = 3,
                                      frac_site_with_ur = 0.2, ur_rate = 0.1)

df_visit2$study_roche <- "B"

df_visit <- bind_rows(df_visit1, df_visit2)

df_site <- site_aggr(df_visit)

df_config <- get_pat_pool_config(df_visit, df_site)

df_config

#> # A tibble: 8 x 6
#> # Groups:   study_roche, site_number [8]
#>   study_roche site_number visit_med75 n_pat_with_med75 n_pat_study pat_pool     
#>   <chr>       <chr>             <dbl>            <dbl>       <dbl> <list>       
#> 1 A           S0001                16               20          68 <tibble [88 …
#> 2 A           S0002                17               17          65 <tibble [82 …
#> 3 A           S0003                15               18          72 <tibble [90 …
#> 4 A           S0004                14               18          76 <tibble [94 …
#> 5 A           S0005                15               18          72 <tibble [90 …
#> 6 B           S0001                15              297         597 <tibble [894…
#> 7 B           S0002                15              296         598 <tibble [894…
#> 8 B           S0003                15              301         593 <tibble [894…