Get Portfolio Event Rates Calculates mean event rates per study and visit in a df_visit simaerep input dataframe.

Get Portfolio Event Rates Calculates mean event rates per study and visit in a df_visit simaerep input dataframe.

Usage

get_portf_event_rates(df_visit, check = TRUE, anonymize = TRUE, pad_width = 4)

Arguments

df_visit

input dataframe with columns study_id, site_id, patient_id, visit, n_events. Can also be a lazy database table.

check

logical, perform standard checks on df_visit, Default: TRUE

anonymize

logical, Default: TRUE

pad_width

padding width for newly created IDs, Default: 4

Examples

df_visit1 <- sim_test_data_study(n_pat = 100, n_sites = 10,
                                 ratio_out = 0.4, factor_event_rate = - 0.6,
                                 study_id = "A")

df_visit2 <- sim_test_data_study(n_pat = 100, n_sites = 10,
                                 ratio_out = 0.2, factor_event_rate = - 0.1,
                                 study_id = "B")

df_visit <- dplyr::bind_rows(df_visit1, df_visit2)

get_portf_event_rates(df_visit)
#> # A tibble: 59 × 4
#>    study_id visit event_rate n_pat
#>    <chr>    <dbl>      <dbl> <int>
#>  1 0001         1      0       100
#>  2 0001         2      1.35    100
#>  3 0001         3      1.5     100
#>  4 0001         4      1.57    100
#>  5 0001         5      0.95    100
#>  6 0001         6      0.87    100
#>  7 0001         7      0.57    100
#>  8 0001         8      0.32    100
#>  9 0001         9      0.263    99
#> 10 0001        10      0.182    99
#> # ℹ 49 more rows

# \donttest{
# Database example
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = ":memory:")
dplyr::copy_to(con, df_visit, "visit")
tbl_visit <- dplyr::tbl(con, "visit")
get_portf_event_rates(tbl_visit)
#> # A tibble: 59 × 4
#>    study_id visit event_rate n_pat
#>    <chr>    <int>      <dbl> <dbl>
#>  1 0001         8     0.32     100
#>  2 0001         9     0.263     99
#>  3 0001        14     0.0421    95
#>  4 0001        20     0.0667    45
#>  5 0002         5     1.47     100
#>  6 0002        24     0.2       15
#>  7 0002        25     0.0909    11
#>  8 0001        28     0          1
#>  9 0001         5     0.95     100
#> 10 0001        24     0         13
#> # ℹ 49 more rows
DBI::dbDisconnect(con)
# }