simulate patients and events for sites supports constant and non-constant event rates

simulate patients and events for sites supports constant and non-constant event rates

Usage

sim_pat(vs_max, vs_sd, is_out, event_rates, event_names, factor_event_rate)