Evaluate sites.

Correct under-reporting probabilities by the expected number of false positives (fp). This has been deprecated in favor of more conventional methods available via p.adjust.

eval_sites_deprecated(df_sim_sites, r_sim_sites)

Arguments

df_sim_sites

dataframe generated by sim_sites()

r_sim_sites

integer, number of repeats for bootstrap resampling for site simulation, needed for zero probability correction for fp calculation, Default: 1000

Value

dataframe with the following columns:

study_id

study identification

site_number

site identification

visit_med75

median(max(visit)) * 0.75

mean_ae_site_med75

mean AE at visit_med75 site level

mean_ae_study_med75

mean AE at visit_med75 study level

pval

p-value as returned by poisson.test

prob_low

bootstrapped probability for having mean_ae_site_med75 or lower

n_site

number of study sites

pval_n_detected

sites with the same p-value or lower

pval_fp

expected number of fp, pval * n_site

pval_p_vs_fp_ratio

odds under-reporting as p/fp, poisson.test (use as benchmark)

pval_prob_ur

probability under-reporting as 1 - fp/p, poisson.test (use as benchmark)

prob_low_n_detected

sites with same bootstrapped probability or lower

prob_low_fp

expected number of fp, prob_lower * n_site

prob_low_p_vs_fp_ratio

odds under-reporting as p/fp, bootstrapped (use)

prob_low_prob_ur

probability under-reporting as 1 - fp/p, bootstrapped (use)

Details

If by chance expected number of false positives (fp) is greater than the total number of positives (p) we set p_vs_fp_ratio = 1 and prob_ur = 0.

See also

Examples

df_visit <- sim_test_data_study(n_pat = 100, n_sites = 5,
    frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit$study_id <- "A"
df_site <- site_aggr(df_visit)

df_sim_sites <- sim_sites(df_site, df_visit, r = 100)

df_eval <- eval_sites_deprecated(df_sim_sites, r_sim_sites = 100)
df_eval
#> # A tibble: 5 × 19
#>   study…¹ site_…² n_pat n_pat…³ visit…⁴ mean_…⁵ mean_…⁶ n_pat…⁷     pval prob_…⁸
#>   <chr>   <chr>   <int>   <int>   <dbl>   <dbl>   <dbl>   <int>    <dbl>   <dbl>
#> 1 A       S0001      20      20      16    3.25    7.18      68 5.77e-11       0
#> 2 A       S0002      20      17      18    3.88    7.81      58 1.20e- 8       0
#> 3 A       S0003      20      18      17   10       5.78      64 1   e+ 0       1
#> 4 A       S0004      20      18      14    7.39    5.01      76 1   e+ 0       1
#> 5 A       S0005      20      18      15    7.28    5.51      72 1   e+ 0       1
#> # … with 9 more variables: n_site <int>, pval_n_detected <int>, pval_fp <dbl>,
#> #   pval_p_vs_fp_ratio <dbl>, pval_prob_ur <dbl>, prob_low_n_detected <int>,
#> #   prob_low_fp <dbl>, prob_low_p_vs_fp_ratio <dbl>, prob_low_prob_ur <dbl>,
#> #   and abbreviated variable names ¹​study_id, ²​site_number, ³​n_pat_with_med75,
#> #   ⁴​visit_med75, ⁵​mean_ae_site_med75, ⁶​mean_ae_study_med75,
#> #   ⁷​n_pat_with_med75_study, ⁸​prob_low