simulate studies

simulate pvalues and prob_lower values for testing

sim_studies(
  df_visit,
  df_site,
  r = r,
  ttest = TRUE,
  prob_lower = TRUE,
  r_prob_lower = 1000,
  parallel = FALSE,
  keep_ae = FALSE,
  min_n_pat_with_med75 = 1,
  studies = NULL,
  .progress = TRUE
)

Arguments

df_visit
dataframe

df_site
dataframe

r
integer, denotes number of simulations, Default: 1000

ttest
logical, calculates ttest pvalue, Default: T

prob_lower
logical, calculates probability for getting a lower value, Default: F

r_prob_lower
nteger, denotes number of simulations for prob_lower value calculation,, Default: 1000

parallel
logical, toggles parallel processing on and of, Default: F

keep_ae
logical, keep ae numbers in output dataframe memory increase roughly 30 percent, Default: F

min_n_pat_with_med75
integer, min number of patients with med75 at site to simulate, Default: 1

studies
vector with study names, Default: NULL

.progress
logical, show progress bar

Value

dataframe

Details

adds column with simulated probabilities for exqual or lower mean_ae at visit_med75

Examples

df_visit1 <- boot_sim_test_data_study(n_pat = 100, n_sites = 5,
                                      frac_site_with_ur = 0.4, ur_rate = 0.6)

df_visit1$study_roche <- "A"

df_visit2 <- boot_sim_test_data_study(n_pat = 1000, n_sites = 3,
                                      frac_site_with_ur = 0.2, ur_rate = 0.1)

df_visit2$study_roche <- "B"

df_visit <- bind_rows(df_visit1, df_visit2)

df_site <- site_aggr(df_visit)

sim_studies(ungroup(df_visit), ungroup(df_site), r = 3, keep_ae = TRUE)

#> # A tibble: 24 x 10
#>        r study_roche site_number visit_med75 n_pat_with_med75 n_pat_study
#>    <dbl> <chr>       <chr>             <dbl>            <dbl>       <dbl>
#>  1     1 A           S0001                16               20          68
#>  2     1 A           S0002                17               17          64
#>  3     1 A           S0003                15               18          72
#>  4     1 A           S0004                14               18          76
#>  5     1 A           S0005                15               18          72
#>  6     1 B           S0001                15              297         597
#>  7     1 B           S0002                15              296         598
#>  8     1 B           S0003                15              301         593
#>  9     2 A           S0001                16               20          68
#> 10     2 A           S0002                17               17          64
#> # … with 14 more rows, and 4 more variables: n_ae_site <chr>, n_ae_study <chr>,
#> #   pval <dbl>, prob_low <dbl>