simulate in dataframe

simulate in dataframe

simaerep_inframe(
  df_visit,
  r = 1000,
  under_only = FALSE,
  visit_med75 = FALSE,
  check = TRUE,
  param_site_aggr = list(method = "med75_adj", min_pat_pool = 0.2),
  param_eval_sites = list(method = "BH"),
  env = parent.frame()
)

Arguments

df_visit

Data frame with columns: study_id, site_number, patnum, visit, n_ae.

r

Integer or tbl_object, number of repetitions for bootstrap simulation. Pass a tbl object referring to a table with one column and as many rows as desired repetitions. Default: 1000.

under_only

Logical, compute under-reporting probabilities only. Supersedes under_only parameter passed to eval_sites() and sim_sites(). Default: TRUE.

visit_med75

Logical, should evaluation point visit_med75 be used. Default: TRUE.

check

Logical, perform data check and attempt repair with check_df_visit(). Computationally expensive on large data sets. Default: TRUE.

param_site_aggr

List of parameters passed to site_aggr(). Default: list(method = "med75_adj", min_pat_pool = 0.2).

param_eval_sites

List of parameters passed to eval_sites(). Default: list(method = "BH").

env

Optional, provide environment of original visit data. Default: parent.frame().

Examples

df_visit <- sim_test_data_study(
 n_pat = 100,
 n_sites = 5,
 frac_site_with_ur = 0.4,
 ur_rate = 0.6
)
df_visit$study_id <- "A"

simaerep_inframe(df_visit)
#> simaerep object:
#> Check aerep$df_eval prob_low_prob_ur column for under-reporting probabililty.
#> Plot results using plot() generic.
simaerep_inframe(df_visit, visit_med75 = TRUE)$df_eval
#> # A tibble: 5 × 12
#>   study_id site_number events_per_visit_site events visits n_pat prob_low
#>   <chr>    <chr>                       <dbl>  <dbl>  <dbl> <int>    <dbl>
#> 1 A        S0001                       0.261     61    234    18    0.015
#> 2 A        S0002                       0.201     61    304    19    0    
#> 3 A        S0003                       0.426    116    272    17    0.925
#> 4 A        S0004                       0.489    130    266    19    0.957
#> 5 A        S0005                       0.481    139    289    17    0.522
#> # ℹ 5 more variables: events_per_visit_study <dbl>, prob_low_adj <dbl>,
#> #   prob_low_prob_ur <dbl>, n_pat_with_med75 <dbl>, visit_med75 <dbl>
# \donttest{
# Database
con <- DBI::dbConnect(duckdb::duckdb(), dbdir = ":memory:")
df_r <- tibble::tibble(rep = seq(1, 1000))

dplyr::copy_to(con, df_visit, "visit")
dplyr::copy_to(con, df_r, "r")

tbl_visit <- dplyr::tbl(con, "visit")
tbl_r <- dplyr::tbl(con, "r")

simaerep_inframe(tbl_visit, r = tbl_r)$df_eval
#> # Source:     SQL [5 x 10]
#> # Database:   DuckDB v1.0.0 [koneswab@Darwin 23.4.0:R 4.4.1/:memory:]
#> # Ordered by: study_id, site_number
#>   study_id site_number events_per_visit_site events visits n_pat prob_low
#>   <chr>    <chr>                       <dbl>  <dbl>  <dbl> <dbl>    <dbl>
#> 1 A        S0003                       0.433    165    381    20    0.827
#> 2 A        S0001                       0.239     84    352    20    0    
#> 3 A        S0002                       0.210     83    396    20    0    
#> 4 A        S0004                       0.480    184    383    20    0.971
#> 5 A        S0005                       0.470    181    385    20    0.949
#> # ℹ 3 more variables: events_per_visit_study <dbl>, prob_low_adj <dbl>,
#> #   prob_low_prob_ur <dbl>
simaerep_inframe(tbl_visit, r = tbl_r, visit_med75 = TRUE)$df_eval
#> # Source:     SQL [5 x 12]
#> # Database:   DuckDB v1.0.0 [koneswab@Darwin 23.4.0:R 4.4.1/:memory:]
#> # Ordered by: study_id, site_number
#>   study_id site_number events_per_visit_site events visits n_pat prob_low
#>   <chr>    <chr>                       <dbl>  <dbl>  <dbl> <dbl>    <dbl>
#> 1 A        S0001                       0.261     61    234    18    0.013
#> 2 A        S0003                       0.426    116    272    17    0.917
#> 3 A        S0004                       0.489    130    266    19    0.963
#> 4 A        S0002                       0.201     61    304    19    0    
#> 5 A        S0005                       0.481    139    289    17    0.533
#> # ℹ 5 more variables: events_per_visit_study <dbl>, prob_low_adj <dbl>,
#> #   prob_low_prob_ur <dbl>, n_pat_with_med75 <dbl>, visit_med75 <int>

DBI::dbDisconnect(con)
# }